requirements.txt
pyyaml==6.0.1
spotipy==2.23.0
langchain==0.0.315
colorama==0.4.6
tiktoken==0.5.1
openai==0.28.1